Split backprop into input and filter
